requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
# mysqlclient==2.2.4  # Drivers para MySQL se comenta para evitar instalacion ya que se requiere tenerlo instalado aparte para poder trabajr de mejor manera
pymysql==1.1.1  # Alternativa pura de Python en manejo de MySQL
cryptography==41.0.7  # Requerido para autenticación MySQL 8.0
python-dotenv==1.0.1
gunicorn==22.0.0
Flask-Migrate==4.0.4
PyJWT==2.8.0  # JWT para manejo de sesiones
flask-cors==4.0.0  # CORS para APIs
requests==2.31.0  # Para tests HTTP
